# src/controllers/controllers.js
import createHttpError from 'http-errors';
import fs from 'node:fs/promises';
import path from 'node:path';
import {
  changeContact,
  createContact,
  deleteContact,
  getAllContacts,
  getContactById,
} from '../services/getContact.js';
import { parsePaginationParams } from '../utils/parsePaginationParams.js';
import { parseSortParams } from '../utils/parseSortParamas.js';
import { parseFilterParams } from '../utils/parseFilterParams.js';
import { env } from '../utils/env.js';
import { uploadToCloudinary } from '../utils/uoloadtoCloudinary.js';

export async function getContactsController(req, res, next) {
  console.log({ 'Користувач цей': req.user });

  const { page, perPage } = parsePaginationParams(req.query);
  const { sortBy, sortOrder } = parseSortParams(req.query);
  const filter = parseFilterParams(req.query);

  console.log('Filter:', filter);
  console.log('User ID:', req.user.id);
  const contacts = await getAllContacts({
    page,
    perPage,
    sortBy,
    sortOrder,
    filter,
    userId: req.user.id,
  });
  res.status(200).json({
    status: 200,
    message: 'Successfully found contacts!',
    data: contacts,
  });
}

export async function getContactController(req, res) {
  const { contactId } = req.params;

  const contact = await getContactById(contactId);
  if (!contact) {
    throw createHttpError(404, 'Contact not found');
  }

  if (contact.userId.toString() !== req.user._id.toString()) {
    throw createHttpError.Forbidden('Contact is forbidden!');
  }
  res.status(200).json({
    status: 200,
    message: `Successfully found contact with id ${contactId}!`,
    data: contact,
  });
}

export async function createContactController(req, res) {
  let photo = null;

  if (req.file) {
    if (env('ENABLE_CLOUDINARY') === 'true') {
      const resultPhoto = await uploadToCloudinary(req.file.path);
      await fs.unlink(req.file.path);

      photo = resultPhoto.secure_url;
    } else {
      //для зміни локації додавання фото
      await fs.rename(
        req.file.path,
        path.resolve('src', 'public/photo', req.file.filename),
      );
      photo = `http://localhost:5000/photo/${req.file.filename}`;
    }
  }

  const contact = {
    name: req.body.name,
    phoneNumber: req.body.phoneNumber,
    email: req.body.email,
    userId: req.user._id,
    photo,
  };

  const result = await createContact(contact);
  res.status(201).json({
    status: 201,
    message: 'Successfully created a contact!',
    data: result,
  });
}

export async function deleteContactController(req, res) {
  const { contactId } = req.params;
  const userId = req.user._id;

  const contact = await deleteContact(contactId, userId);

  if (!contact || contact.userId.toString() !== req.user._id.toString()) {
    throw createHttpError(
      404,
      'Contact not found or you are not authorized to delete this contact!',
    );
  }

  res.status(204).end();
}

export async function changeContactController(req, res) {
  const { contactId } = req.params;
  const userId = req.user._id;
  const updateData = { ...req.body };

  let photo = null;

  if (req.file) {
    if (env('ENABLE_CLOUDINARY') === 'true') {
      const resultPhoto = await uploadToCloudinary(req.file.path);
      await fs.unlink(req.file.path);

      photo = resultPhoto.secure_url;
    } else {
      const newPath = path.resolve('src', 'public/photo', req.file.filename);
      //для зміни локації додавання фото
      await fs.rename(req.file.path, newPath);
      photo = `http://localhost:5000/photo/${req.file.filename}`;
    }
    //додаємо оновлення вже з фото
    updateData.photo = photo;
  }

  const updatedContact = await changeContact(contactId, userId, updateData);
  if (!updatedContact) {
    throw createHttpError(
      404,
      'Contact not found or you are not authorized to delete this contact!',
    );
  }
  res.json({
    status: 200,
    message: 'Contact changes successfully',
    data: updatedContact,
  });
}

/* export async function updateContactController(req, res) {
  const { contactId } = req.params;

  const contact = {
    name: req.body.name,
    phoneNumber: req.body.phoneNumber,
    email: req.body.email,
    isFavorite: req.body.isFavorite,
    contactType: req.body.contactType,
  };

  console.log(contact);

  const result = await updateContact(contactId, contact);

  if (result === null) {
    throw createHttpError(404, "Contact not found");
  }

  res.json({
    status: 200,
    message: "Contact updated successfully",
    data: result,
  });
} */